requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.10.1
boto3==1.26.106
dbt-core==1.8.1
dbt-snowflake==1.8.1
ydata-profiling==4.16.1
